SpotMoodBot/requirements.txt
botbuilder-integration-aiohttp>=4.13.0
recognizers-text>=1.0.2a1